Updates to match holdings changes in Argon.
