+++ .rubocop.yml
@@ -21,3 +21,7 @@ Metrics/BlockLength:
   Exclude:
     - 'spec/**/*'
     - 'app/controllers/catalog_controller.rb'
+
+Metrics/ModuleLength:
+  Exclude:
+    - 'app/helpers/trln_argon_helper.rb'
