+++ Gemfile
@@ -61,7 +61,9 @@ gem 'tzinfo-data', platforms: %i[mingw mswin x64_mingw jruby]
 gem 'addressable', '~> 2.5'
 gem 'blacklight', '~> 6.7'
 gem 'lcsort', '~> 0.9'
-gem 'trln_argon', git: 'https://github.com/trln/trln_argon'
+gem 'trln_argon',
+    git: 'https://github.com/trln/trln_argon',
+    ref: 'b0195c2bb8ff97c0079584e981f1e29ee790e542'
 
 group :development, :test do
   gem 'rails-controller-testing'
