+++ spec/fixtures/files/holdings/no_items_with_holdings_has_summary.yml
@@ -1,0 +1,13 @@
+---
+PERKN:
+  PEI:
+    items: []
+    holdings:
+    - summary: 'Available at address: http://wwwrvares.er.usgs.gov/nawqa/WRI94-4001.html'
+DOCS:
+  PGS:
+    items: []
+    holdings:
+    - notes:
+      - Not distributed in print; see online version
+      call_no: I 19.42/4:94-4001
